- Update PDL type - Switch C-Style cast to `static_cast` - Revert Version.hpp

+++ examples/cpp/SampleCpp/main.cpp
@@ -362,8 +362,8 @@ int main()
         event.SetProperty("random", rand());
         event.SetProperty("secret", 5.6872);
         event.SetProperty("seq", (uint64_t)i); 
-        event.SetProperty(COMMONFIELDS_EVENT_PRIVTAGS, (int64_t)(i+1) );
-        event.SetProperty(COMMONFIELDS_EVENT_PRIVLEVEL, (int64_t)(i+1) );
+        event.SetProperty(COMMONFIELDS_EVENT_PRIVTAGS, static_cast<int64_t>(i+1) );
+        event.SetProperty(COMMONFIELDS_EVENT_PRIVLEVEL, static_cast<uint32_t>(i + 1));
         event.SetLatency(latency); 
         logger->LogEvent(event);
 
@@ -373,8 +373,8 @@ int main()
                 { "random", rand() },
                 { "secret", 5.6872 },
                 { "seq", (uint64_t)i },
-                { COMMONFIELDS_EVENT_PRIVTAGS, (int64_t)(i + 1) }
-                { COMMONFIELDS_EVENT_PRIVLEVEL, (int64_t)(i + 1) }
+                {COMMONFIELDS_EVENT_PRIVTAGS, static_cast<int64_t>(i + 1)},
+                {COMMONFIELDS_EVENT_PRIVLEVEL, static_cast<uint32_t>(i + 1)}
             });
         logger->LogEvent(event2);
     }
